pyproject.toml
[tool.poetry]
name = "prompt-compiler"
version = "0.1.0"
description = "A compiler that generates code from prompt files"
authors = ["Jeonghan Joo <joo3436@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0"
jinja2 = "^3.1.2"
openai = "^1.12.0"
anthropic = "^0.18.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.1"
isort = "^5.13.2"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
prompt-compiler = "prompt_compiler.cli:main"
